Adjust the default client-identifier
It looks like ActiveMQ is quite happy with the empty string and makes a
broker generated id, whereas mosquitto doesn't appear to like it
regardless of the setting of the "clean-session" flag. So this is just
to stop surprising people like me.

Also set the remaining payload length in subscribe because it seems that
mosquitto doesn't like that not being set, ActiveMQ being perfectly happy.